sidecar/requirements.txt
# WisDev Agent OS Python sidecar dependencies
# Optional worker service for heavy workloads

# FastAPI framework
fastapi==0.133.0
uvicorn[standard]==0.41.0
python-multipart==0.0.22

# Google Cloud & GenAI SDK
google-genai==1.62.0
google-cloud-aiplatform==1.138.0
google-auth==2.47.0

# Agent harness
deepagents==0.5.1
langchain==1.2.15

# Vector Database (Pinecone)
pinecone-client==6.0.0

# Redis Cache (Upstash)
upstash-redis==1.6.0
redis==7.2.0

# HTTP clients
httpx==0.28.1
aiohttp==3.13.3
grpcio==1.78.0
grpcio-tools==1.78.0
# Keep runtime pins aligned with the checked-in protobuf/gRPC stubs.
protobuf==6.33.5

# Data validation
pydantic==2.12.5
pydantic-settings==2.13.1

# Document processing
docling>=2.3.0
pypdf==6.7.2
python-docx==1.2.0
pymupdf>=1.27.1
# Optional: install google/langextract separately if you want the PDF metadata
# LLM fallback. The sidecar falls back to regex extraction when it is absent.

# Export formats
reportlab==4.4.10
Jinja2==3.1.6
markdown==3.10.2

# SSE streaming
sse-starlette==3.2.0

# Utilities
numpy==2.4.4
scipy==1.16.3
scikit-learn==1.7.2
tiktoken==0.12.0
tenacity==9.1.4
monty>=2026.2.18
rapidfuzz>=3.14.3
rank-bm25==0.2.2
python-dotenv==1.2.1
structlog==25.5.0

# Secret Management
google-cloud-secret-manager==2.27.0

# Rate limiting
slowapi==0.1.9

# ── OpenTelemetry (observability) ──────────────────────────────────────────
# Core SDK + GCP Cloud Trace exporter + FastAPI auto-instrumentation.
opentelemetry-sdk==1.34.1
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-propagator-gcp==1.9.0
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-instrumentation-grpc==0.55b1

# Async Postgres driver (used by dynamic_skill_registry for skill persistence)
asyncpg>=0.30.0

# Testing (dev only)
pytest>=9.0.2
pytest-cov==7.0.0
pytest-asyncio==1.3.0
mypy>=1.18.2
